+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## [Unreleased]
 
 ### Added
+- Context menu with "Save playlist" function for advanced search results
+  [#1256](https://github.com/owncloud/music/issues/1256)
 
 ### Changed
 - Optimization for the scan speed (most significant on slow DB like SQLite)